requirements.txt
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
pandas>=2.0.0
pyarrow>=12.0.0
db-dtypes>=1.2.0
python-dotenv>=1.0.0

